Switch license to MIT + APLv2
According to the GPL FAQ, LGPL forces you to provide all the objects or
sources required to relink a binary, when using a static library.
In Rust, the default is to use static linking, and providing all objects
is not really suitable.
To allow a wider use of this crate, switch to MIT + APLv2.

[1] https://www.gnu.org/licenses/gpl-faq.html#LGPLStaticVsDynamic
ease enter the commit message for your changes. Lines starting